volta_core

Module error

Modules

kind 🔒

reporter 🔒

Structs

Inner 🔒

VoltaError

Error type for Volta

Enums

ErrorKind

ExitCode

Exit codes supported by Volta Errors

Traits

Context

Trait providing the with_context method to easily convert any Result error into a VoltaError

Functions

report_error

Report an error, both to the console and to error logs

Type Aliases

Fallible