Enum volta_core::error::ErrorKind

pub enum ErrorKind {

    BinaryAlreadyInstalled {
        bin_name: String,
        existing_package: String,
        new_package: String,
    },
    BinaryExecError,
    BinaryNotFound {
        name: String,
    },
    BuildPathError,
    BypassError {
        command: String,
    },
    CannotFetchPackage {
        package: String,
    },
    CannotPinPackage {
        package: String,
    },
    CompletionsOutFileError {
        path: PathBuf,
    },
    ContainingDirError {
        path: PathBuf,
    },
    CouldNotDetermineTool,
    CouldNotStartMigration,
    CreateDirError {
        dir: PathBuf,
    },
    CreateLayoutFileError {
        file: PathBuf,
    },
    CreateSharedLinkError {
        name: String,
    },
    CreateTempDirError {
        in_dir: PathBuf,
    },
    CreateTempFileError {
        in_dir: PathBuf,
    },
    CurrentDirError,
    DeleteDirectoryError {
        directory: PathBuf,
    },
    DeleteFileError {
        file: PathBuf,
    },
    DeprecatedCommandError {
        command: String,
        advice: String,
    },
    DownloadToolNetworkError {
        tool: Spec,
        from_url: String,
    },
    ExecuteHookError {
        command: String,
    },
    ExtensionCycleError {
        paths: Vec<PathBuf>,
        duplicate: PathBuf,
    },
    ExtensionPathError {
        path: PathBuf,
    },
    HookCommandFailed {
        command: String,
    },
    HookMultipleFieldsSpecified,
    HookNoFieldsSpecified,
    HookPathError {
        command: String,
    },
    InstalledPackageNameError,
    InvalidHookCommand {
        command: String,
    },
    InvalidHookOutput {
        command: String,
    },
    InvalidInvocation {
        action: String,
        name: String,
        version: String,
    },
    InvalidInvocationOfBareVersion {
        action: String,
        version: String,
    },
    InvalidRegistryFormat {
        format: String,
    },
    InvalidToolName {
        name: String,
        errors: Vec<String>,
    },
    LockAcquireError,
    NoBundledNpm {
        command: String,
    },
    NoCommandLinePnpm,
    NoCommandLineYarn,
    NoDefaultNodeVersion {
        tool: String,
    },
    NodeVersionNotFound {
        matching: String,
    },
    NoHomeEnvironmentVar,
    NoInstallDir,
    NoLocalDataDir,
    NoPinnedNodeVersion {
        tool: String,
    },
    NoPlatform,
    NoProjectNodeInManifest,
    NoProjectYarn,
    NoProjectPnpm,
    NoShellProfile {
        env_profile: String,
        bin_dir: PathBuf,
    },
    NotInPackage,
    NoDefaultYarn,
    NoDefaultPnpm,
    NpmLinkMissingPackage {
        package: String,
    },
    NpmLinkWrongManager {
        package: String,
    },
    NpmVersionNotFound {
        matching: String,
    },
    NpxNotAvailable {
        version: String,
    },
    PackageInstallFailed {
        package: String,
    },
    PackageManifestParseError {
        package: String,
    },
    PackageManifestReadError {
        package: String,
    },
    PackageNotFound {
        package: String,
    },
    PackageParseError {
        file: PathBuf,
    },
    PackageReadError {
        file: PathBuf,
    },
    PackageUnpackError,
    PackageWriteError {
        file: PathBuf,
    },
    ParseBinConfigError,
    ParseHooksError {
        file: PathBuf,
    },
    ParseNodeIndexCacheError,
    ParseNodeIndexError {
        from_url: String,
    },
    ParseNodeIndexExpiryError,
    ParseNpmManifestError,
    ParsePackageConfigError,
    ParsePlatformError,
    ParseToolSpecError {
        tool_spec: String,
    },
    PersistInventoryError {
        tool: String,
    },
    PnpmVersionNotFound {
        matching: String,
    },
    ProjectLocalBinaryExecError {
        command: String,
    },
    ProjectLocalBinaryNotFound {
        command: String,
    },
    PublishHookBothUrlAndBin,
    PublishHookNeitherUrlNorBin,
    ReadBinConfigDirError {
        dir: PathBuf,
    },
    ReadBinConfigError {
        file: PathBuf,
    },
    ReadDefaultNpmError {
        file: PathBuf,
    },
    ReadDirError {
        dir: PathBuf,
    },
    ReadHooksError {
        file: PathBuf,
    },
    ReadNodeIndexCacheError {
        file: PathBuf,
    },
    ReadNodeIndexExpiryError {
        file: PathBuf,
    },
    ReadNpmManifestError,
    ReadPackageConfigError {
        file: PathBuf,
    },
    ReadPlatformError {
        file: PathBuf,
    },
    RegistryFetchError {
        tool: String,
        from_url: String,
    },
    RunShimDirectly,
    SetToolExecutable {
        tool: String,
    },
    SetupToolImageError {
        tool: String,
        version: String,
        dir: PathBuf,
    },
    ShimCreateError {
        name: String,
    },
    ShimRemoveError {
        name: String,
    },
    StringifyBinConfigError,
    StringifyPackageConfigError,
    StringifyPlatformError,
    Unimplemented {
        feature: String,
    },
    UnpackArchiveError {
        tool: String,
        version: String,
    },
    UpgradePackageNotFound {
        package: String,
        manager: PackageManager,
    },
    UpgradePackageWrongManager {
        package: String,
        manager: PackageManager,
    },
    VersionParseError {
        version: String,
    },
    WriteBinConfigError {
        file: PathBuf,
    },
    WriteDefaultNpmError {
        file: PathBuf,
    },
    WriteLauncherError {
        tool: String,
    },
    WriteNodeIndexCacheError {
        file: PathBuf,
    },
    WriteNodeIndexExpiryError {
        file: PathBuf,
    },
    WritePackageConfigError {
        file: PathBuf,
    },
    WritePlatformError {
        file: PathBuf,
    },
    Yarn2NotSupported,
    YarnLatestFetchError {
        from_url: String,
    },
    YarnVersionNotFound {
        matching: String,
    },

}

Variants

BinaryAlreadyInstalled

Fields

bin_name: String

existing_package: String

new_package: String

Thrown when package tries to install a binary that is already installed.

BinaryExecError

Thrown when executing an external binary fails

BinaryNotFound

Fields

name: String

Thrown when a binary could not be found in the local inventory

BuildPathError

Thrown when building the virtual environment path fails

BypassError

Fields

command: String

Thrown when unable to launch a command with VOLTA_BYPASS set

CannotFetchPackage

Fields

package: String

Thrown when a user tries to volta fetch something other than node/yarn/npm.

CannotPinPackage

Fields

package: String

Thrown when a user tries to volta pin something other than node/yarn/npm.

CompletionsOutFileError

Fields

path: PathBuf

Thrown when the Completions out-dir is not a directory

ContainingDirError

Fields

path: PathBuf

Thrown when the containing directory could not be determined

CouldNotDetermineTool

CouldNotStartMigration

Thrown when unable to start the migration executable

CreateDirError

Fields

dir: PathBuf

CreateLayoutFileError

Fields

file: PathBuf

Thrown when unable to create the layout file

CreateSharedLinkError

Fields

name: String

Thrown when unable to create a link to the shared global library directory

CreateTempDirError

Fields

in_dir: PathBuf

Thrown when creating a temporary directory fails

CreateTempFileError

Fields

in_dir: PathBuf

Thrown when creating a temporary file fails

CurrentDirError

DeleteDirectoryError

Fields

directory: PathBuf

Thrown when deleting a directory fails

DeleteFileError

Fields

file: PathBuf

Thrown when deleting a file fails

DeprecatedCommandError

Fields

command: String

advice: String

DownloadToolNetworkError

Fields

tool: Spec

from_url: String

ExecuteHookError

Fields

command: String

Thrown when unable to execute a hook command

ExtensionCycleError

Fields

paths: Vec<PathBuf>

duplicate: PathBuf

Thrown when volta.extends keys result in an infinite cycle

ExtensionPathError

Fields

path: PathBuf

Thrown when determining the path to an extension manifest fails

HookCommandFailed

Fields

command: String

Thrown when a hook command returns a non-zero exit code

HookMultipleFieldsSpecified

Thrown when a hook contains multiple fields (prefix, template, or bin)

HookNoFieldsSpecified

Thrown when a hook doesn’t contain any of the known fields (prefix, template, or bin)

HookPathError

Fields

command: String

Thrown when determining the path to a hook fails

InstalledPackageNameError

Thrown when determining the name of a newly-installed package fails

InvalidHookCommand

Fields

command: String

InvalidHookOutput

Fields

command: String

Thrown when output from a hook command could not be read

InvalidInvocation

Fields

action: String

name: String

version: String

Thrown when a user does e.g. volta install node 12 instead of volta install node@12.

InvalidInvocationOfBareVersion

Fields

action: String

version: String

Thrown when a user does e.g. volta install 12 instead of volta install node@12.

InvalidRegistryFormat

Fields

format: String

Thrown when a format other than “npm” or “github” is given for yarn.index in the hooks

InvalidToolName

Fields

name: String

errors: Vec<String>

Thrown when a tool name is invalid per npm’s rules.

LockAcquireError

Thrown when unable to acquire a lock on the Volta directory

NoBundledNpm

Fields

command: String

Thrown when pinning or installing npm@bundled and couldn’t detect the bundled version

NoCommandLinePnpm

Thrown when pnpm is not set at the command-line

NoCommandLineYarn

Thrown when Yarn is not set at the command-line

NoDefaultNodeVersion

Fields

tool: String

Thrown when a user tries to install a Yarn or npm version before installing a Node version.

NodeVersionNotFound

Fields

matching: String

Thrown when there is no Node version matching a requested semver specifier.

NoHomeEnvironmentVar

NoInstallDir

Thrown when the install dir could not be determined

NoLocalDataDir

NoPinnedNodeVersion

Fields

tool: String

Thrown when a user tries to pin a npm, pnpm, or Yarn version before pinning a Node version.

NoPlatform

Thrown when the platform (Node version) could not be determined

NoProjectNodeInManifest

Thrown when parsing the project manifest and there is a "volta" key without Node

NoProjectYarn

Thrown when Yarn is not set in a project

NoProjectPnpm

Thrown when pnpm is not set in a project

NoShellProfile

Fields

env_profile: String

bin_dir: PathBuf

Thrown when no shell profiles could be found

NotInPackage

Thrown when the user tries to pin Node or Yarn versions outside of a package.

NoDefaultYarn

Thrown when default Yarn is not set

NoDefaultPnpm

Thrown when default pnpm is not set

NpmLinkMissingPackage

Fields

package: String

Thrown when npm link is called with a package that isn’t available

NpmLinkWrongManager

Fields

package: String

Thrown when npm link is called with a package that was not installed / linked with npm

NpmVersionNotFound

Fields

matching: String

Thrown when there is no npm version matching the requested Semver/Tag

NpxNotAvailable

Fields

version: String

PackageInstallFailed

Fields

package: String

Thrown when the command to install a global package is not successful

PackageManifestParseError

Fields

package: String

Thrown when parsing the package manifest fails

PackageManifestReadError

Fields

package: String

Thrown when reading the package manifest fails

PackageNotFound

Fields

package: String

Thrown when a specified package could not be found on the npm registry

PackageParseError

Fields

file: PathBuf

Thrown when parsing a package manifest fails

PackageReadError

Fields

file: PathBuf

Thrown when reading a package manifest fails

PackageUnpackError

Thrown when a package has been unpacked but is not formed correctly.

PackageWriteError

Fields

file: PathBuf

Thrown when writing a package manifest fails

ParseBinConfigError

Thrown when unable to parse a bin config file

ParseHooksError

Fields

file: PathBuf

Thrown when unable to parse a hooks.json file

ParseNodeIndexCacheError

Thrown when unable to parse the node index cache

ParseNodeIndexError

Fields

from_url: String

Thrown when unable to parse the node index

ParseNodeIndexExpiryError

Thrown when unable to parse the node index cache expiration

ParseNpmManifestError

Thrown when unable to parse the npm manifest file from a node install

ParsePackageConfigError

Thrown when unable to parse a package configuration

ParsePlatformError

Thrown when unable to parse the platform.json file

ParseToolSpecError

Fields

tool_spec: String

Thrown when unable to parse a tool spec (<tool>[@<version>])

PersistInventoryError

Fields

tool: String

Thrown when persisting an archive to the inventory fails

PnpmVersionNotFound

Fields

matching: String

Thrown when there is no pnpm version matching a requested semver specifier.

ProjectLocalBinaryExecError

Fields

command: String

Thrown when executing a project-local binary fails

ProjectLocalBinaryNotFound

Fields

command: String

Thrown when a project-local binary could not be found

PublishHookBothUrlAndBin

Thrown when a publish hook contains both the url and bin fields

PublishHookNeitherUrlNorBin

Thrown when a publish hook contains neither url nor bin fields

ReadBinConfigDirError

Fields

dir: PathBuf

Thrown when there was an error reading the user bin directory

ReadBinConfigError

Fields

file: PathBuf

Thrown when there was an error reading the config for a binary

ReadDefaultNpmError

Fields

file: PathBuf

Thrown when unable to read the default npm version file

ReadDirError

Fields

dir: PathBuf

Thrown when unable to read the contents of a directory

ReadHooksError

Fields

file: PathBuf

Thrown when there was an error opening a hooks.json file

ReadNodeIndexCacheError

Fields

file: PathBuf

Thrown when there was an error reading the Node Index Cache

ReadNodeIndexExpiryError

Fields

file: PathBuf

Thrown when there was an error reading the Node Index Cache Expiration

ReadNpmManifestError

Thrown when there was an error reading the npm manifest file

ReadPackageConfigError

Fields

file: PathBuf

Thrown when there was an error reading a package configuration file

ReadPlatformError

Fields

file: PathBuf

Thrown when there was an error opening the user platform file

RegistryFetchError

Fields

tool: String

from_url: String

Thrown when the public registry for Node or Yarn could not be downloaded.

RunShimDirectly

Thrown when the shim binary is called directly, not through a symlink

SetToolExecutable

Fields

tool: String

Thrown when there was an error setting a tool to executable

SetupToolImageError

Fields

tool: String

version: String

dir: PathBuf

Thrown when there was an error copying an unpacked tool to the image directory

ShimCreateError

Fields

name: String

Thrown when Volta is unable to create a shim

ShimRemoveError

Fields

name: String

Thrown when Volta is unable to remove a shim

StringifyBinConfigError

Thrown when serializing a bin config to JSON fails

StringifyPackageConfigError

Thrown when serializing a package config to JSON fails

StringifyPlatformError

Thrown when serializing the platform to JSON fails

Unimplemented

Fields

feature: String

Thrown when a given feature has not yet been implemented

UnpackArchiveError

Fields

tool: String

version: String

Thrown when unpacking an archive (tarball or zip) fails

UpgradePackageNotFound

Fields

package: String

manager: PackageManager

Thrown when a package to upgrade was not found

UpgradePackageWrongManager

Fields

package: String

manager: PackageManager

Thrown when a package to upgrade was installed with a different package manager

VersionParseError

Fields

version: String

WriteBinConfigError

Fields

file: PathBuf

Thrown when there was an error writing a bin config file

WriteDefaultNpmError

Fields

file: PathBuf

Thrown when there was an error writing the default npm to file

WriteLauncherError

Fields

tool: String

Thrown when there was an error writing the npm launcher

WriteNodeIndexCacheError

Fields

file: PathBuf

Thrown when there was an error writing the node index cache

WriteNodeIndexExpiryError

Fields

file: PathBuf

Thrown when there was an error writing the node index expiration

WritePackageConfigError

Fields

file: PathBuf

Thrown when there was an error writing a package config

WritePlatformError

Fields

file: PathBuf

Thrown when writing the platform.json file fails

Yarn2NotSupported

Thrown when a user attempts to install a version of Yarn2

YarnLatestFetchError

Fields

from_url: String

Thrown when there is an error fetching the latest version of Yarn

YarnVersionNotFound

Fields

matching: String

Thrown when there is no Yarn version matching a requested semver specifier.

Implementations

impl ErrorKind

pub fn exit_code(&self) -> ExitCode