Enum volta_core::error::kind::ErrorKind

pub enum ErrorKind {

    BinaryAlreadyInstalled {
        bin_name: String,
        existing_package: String,
        new_package: String,
    },
    BinaryExecError,
    BinaryNotFound {
        name: String,
    },
    BuildPathError,
    BypassError {
        command: String,
    },
    CannotFetchPackage {
        package: String,
    },
    CannotPinPackage {
        package: String,
    },
    CompletionsOutFileError {
        path: PathBuf,
    },
    ContainingDirError {
        path: PathBuf,
    },
    CouldNotDetermineTool,
    CouldNotStartMigration,
    CreateDirError {
        dir: PathBuf,
    },
    CreateLayoutFileError {
        file: PathBuf,
    },
    CreateSharedLinkError {
        name: String,
    },
    CreateTempDirError {
        in_dir: PathBuf,
    },
    CreateTempFileError {
        in_dir: PathBuf,
    },
    CurrentDirError,
    DeleteDirectoryError {
        directory: PathBuf,
    },
    DeleteFileError {
        file: PathBuf,
    },
    DeprecatedCommandError {
        command: String,
        advice: String,
    },
    DownloadToolNetworkError {
        tool: Spec,
        from_url: String,
    },
    ExecuteHookError {
        command: String,
    },
    ExtensionCycleError {
        paths: Vec<PathBuf>,
        duplicate: PathBuf,
    },
    ExtensionPathError {
        path: PathBuf,
    },
    HookCommandFailed {
        command: String,
    },
    HookMultipleFieldsSpecified,
    HookNoFieldsSpecified,
    HookPathError {
        command: String,
    },
    InstalledPackageNameError,
    InvalidHookCommand {
        command: String,
    },
    InvalidHookOutput {
        command: String,
    },
    InvalidInvocation {
        action: String,
        name: String,
        version: String,
    },
    InvalidInvocationOfBareVersion {
        action: String,
        version: String,
    },
    InvalidToolName {
        name: String,
        errors: Vec<String>,
    },
    LockAcquireError,
    NoBundledNpm {
        command: String,
    },
    NoCommandLineYarn,
    NoDefaultNodeVersion {
        tool: String,
    },
    NodeVersionNotFound {
        matching: String,
    },
    NoHomeEnvironmentVar,
    NoInstallDir,
    NoLocalDataDir,
    NoPinnedNodeVersion {
        tool: String,
    },
    NoPlatform,
    NoProjectNodeInManifest,
    NoProjectYarn,
    NoShellProfile {
        env_profile: String,
        bin_dir: PathBuf,
    },
    NotInPackage,
    NoDefaultYarn,
    NpmLinkMissingPackage {
        package: String,
    },
    NpmLinkWrongManager {
        package: String,
    },
    NpmVersionNotFound {
        matching: String,
    },
    NpxNotAvailable {
        version: String,
    },
    PackageInstallFailed {
        package: String,
    },
    PackageManifestParseError {
        package: String,
    },
    PackageManifestReadError {
        package: String,
    },
    PackageNotFound {
        package: String,
    },
    PackageParseError {
        file: PathBuf,
    },
    PackageReadError {
        file: PathBuf,
    },
    PackageUnpackError,
    PackageWriteError {
        file: PathBuf,
    },
    ParseBinConfigError,
    ParseHooksError {
        file: PathBuf,
    },
    ParseNodeIndexCacheError,
    ParseNodeIndexError {
        from_url: String,
    },
    ParseNodeIndexExpiryError,
    ParseNpmManifestError,
    ParsePackageConfigError,
    ParsePlatformError,
    ParseToolSpecError {
        tool_spec: String,
    },
    PersistInventoryError {
        tool: String,
    },
    ProjectLocalBinaryExecError {
        command: String,
    },
    ProjectLocalBinaryNotFound {
        command: String,
    },
    PublishHookBothUrlAndBin,
    PublishHookNeitherUrlNorBin,
    ReadBinConfigDirError {
        dir: PathBuf,
    },
    ReadBinConfigError {
        file: PathBuf,
    },
    ReadDefaultNpmError {
        file: PathBuf,
    },
    ReadDirError {
        dir: PathBuf,
    },
    ReadHooksError {
        file: PathBuf,
    },
    ReadNodeIndexCacheError {
        file: PathBuf,
    },
    ReadNodeIndexExpiryError {
        file: PathBuf,
    },
    ReadNpmManifestError,
    ReadPackageConfigError {
        file: PathBuf,
    },
    ReadPlatformError {
        file: PathBuf,
    },
    RegistryFetchError {
        tool: String,
        from_url: String,
    },
    RunShimDirectly,
    SetupToolImageError {
        tool: String,
        version: String,
        dir: PathBuf,
    },
    ShimCreateError {
        name: String,
    },
    ShimRemoveError {
        name: String,
    },
    StringifyBinConfigError,
    StringifyPackageConfigError,
    StringifyPlatformError,
    Unimplemented {
        feature: String,
    },
    UnpackArchiveError {
        tool: String,
        version: String,
    },
    UpgradePackageNotFound {
        package: String,
        manager: PackageManager,
    },
    UpgradePackageWrongManager {
        package: String,
        manager: PackageManager,
    },
    VersionParseError {
        version: String,
    },
    WriteBinConfigError {
        file: PathBuf,
    },
    WriteDefaultNpmError {
        file: PathBuf,
    },
    WriteLauncherError {
        tool: String,
    },
    WriteNodeIndexCacheError {
        file: PathBuf,
    },
    WriteNodeIndexExpiryError {
        file: PathBuf,
    },
    WritePackageConfigError {
        file: PathBuf,
    },
    WritePlatformError {
        file: PathBuf,
    },
    YarnLatestFetchError {
        from_url: String,
    },
    YarnVersionNotFound {
        matching: String,
    },

}

Variants

BinaryAlreadyInstalled

Thrown when package tries to install a binary that is already installed.

Fields of BinaryAlreadyInstalled

bin_name: Stringexisting_package: Stringnew_package: String

BinaryExecError

Thrown when executing an external binary fails

BinaryNotFound

Thrown when a binary could not be found in the local inventory

Fields of BinaryNotFound

name: String

BuildPathError

Thrown when building the virtual environment path fails

BypassError

Thrown when unable to launch a command with VOLTA_BYPASS set

Fields of BypassError

command: String

CannotFetchPackage

Thrown when a user tries to volta fetch something other than node/yarn/npm.

Fields of CannotFetchPackage

package: String

CannotPinPackage

Thrown when a user tries to volta pin something other than node/yarn/npm.

Fields of CannotPinPackage

package: String

CompletionsOutFileError

Thrown when the Completions out-dir is not a directory

Fields of CompletionsOutFileError

path: PathBuf

ContainingDirError

Thrown when the containing directory could not be determined

Fields of ContainingDirError

path: PathBuf

CouldNotDetermineTool

CouldNotStartMigration

Thrown when unable to start the migration executable

CreateDirError

Fields of CreateDirError

dir: PathBuf

CreateLayoutFileError

Thrown when unable to create the layout file

Fields of CreateLayoutFileError

file: PathBuf

CreateSharedLinkError

Thrown when unable to create a link to the shared global library directory

Fields of CreateSharedLinkError

name: String

CreateTempDirError

Thrown when creating a temporary directory fails

Fields of CreateTempDirError

in_dir: PathBuf

CreateTempFileError

Thrown when creating a temporary file fails

Fields of CreateTempFileError

in_dir: PathBuf

CurrentDirError

DeleteDirectoryError

Thrown when deleting a directory fails

Fields of DeleteDirectoryError

directory: PathBuf

DeleteFileError

Thrown when deleting a file fails

Fields of DeleteFileError

file: PathBuf

DeprecatedCommandError

Fields of DeprecatedCommandError

command: Stringadvice: String

DownloadToolNetworkError

Fields of DownloadToolNetworkError

tool: Specfrom_url: String

ExecuteHookError

Thrown when unable to execute a hook command

Fields of ExecuteHookError

command: String

ExtensionCycleError

Thrown when volta.extends keys result in an infinite cycle

Fields of ExtensionCycleError

paths: Vec<PathBuf>duplicate: PathBuf

ExtensionPathError

Thrown when determining the path to an extension manifest fails

Fields of ExtensionPathError

path: PathBuf

HookCommandFailed

Thrown when a hook command returns a non-zero exit code

Fields of HookCommandFailed

command: String

HookMultipleFieldsSpecified

Thrown when a hook contains multiple fields (prefix, template, or bin)

HookNoFieldsSpecified

Thrown when a hook doesn’t contain any of the known fields (prefix, template, or bin)

HookPathError

Thrown when determining the path to a hook fails

Fields of HookPathError

command: String

InstalledPackageNameError

Thrown when determining the name of a newly-installed package fails

InvalidHookCommand

Fields of InvalidHookCommand

command: String

InvalidHookOutput

Thrown when output from a hook command could not be read

Fields of InvalidHookOutput

command: String

InvalidInvocation

Thrown when a user does e.g. volta install node 12 instead of volta install node@12.

Fields of InvalidInvocation

action: Stringname: Stringversion: String

InvalidInvocationOfBareVersion

Thrown when a user does e.g. volta install 12 instead of volta install node@12.

Fields of InvalidInvocationOfBareVersion

action: Stringversion: String

InvalidToolName

Thrown when a tool name is invalid per npm’s rules.

Fields of InvalidToolName

name: Stringerrors: Vec<String>

LockAcquireError

Thrown when unable to acquire a lock on the Volta directory

NoBundledNpm

Thrown when pinning or installing npm@bundled and couldn’t detect the bundled version

Fields of NoBundledNpm

command: String

NoCommandLineYarn

Thrown when Yarn is not set at the command-line

NoDefaultNodeVersion

Thrown when a user tries to install a Yarn or npm version before installing a Node version.

Fields of NoDefaultNodeVersion

tool: String

NodeVersionNotFound

Thrown when there is no Node version matching a requested semver specifier.

Fields of NodeVersionNotFound

matching: String

NoHomeEnvironmentVar

NoInstallDir

Thrown when the install dir could not be determined

NoLocalDataDir

NoPinnedNodeVersion

Thrown when a user tries to pin a Yarn or npm version before pinning a Node version.

Fields of NoPinnedNodeVersion

tool: String

NoPlatform

Thrown when the platform (Node version) could not be determined

NoProjectNodeInManifest

Thrown when parsing the project manifest and there is a "volta" key without Node

NoProjectYarn

Thrown when Yarn is not set in a project

NoShellProfile

Thrown when no shell profiles could be found

Fields of NoShellProfile

env_profile: Stringbin_dir: PathBuf

NotInPackage

Thrown when the user tries to pin Node or Yarn versions outside of a package.

NoDefaultYarn

Thrown when default Yarn is not set

NpmLinkMissingPackage

Thrown when npm link is called with a package that isn’t available

Fields of NpmLinkMissingPackage

package: String

NpmLinkWrongManager

Thrown when npm link is called with a package that was not installed / linked with npm

Fields of NpmLinkWrongManager

package: String

NpmVersionNotFound

Thrown when there is no npm version matching the requested Semver/Tag

Fields of NpmVersionNotFound

matching: String

NpxNotAvailable

Fields of NpxNotAvailable

version: String

PackageInstallFailed

Thrown when the command to install a global package is not successful

Fields of PackageInstallFailed

package: String

PackageManifestParseError

Thrown when parsing the package manifest fails

Fields of PackageManifestParseError

package: String

PackageManifestReadError

Thrown when reading the package manifest fails

Fields of PackageManifestReadError

package: String

PackageNotFound

Thrown when a specified package could not be found on the npm registry

Fields of PackageNotFound

package: String

PackageParseError

Thrown when parsing a package manifest fails

Fields of PackageParseError

file: PathBuf

PackageReadError

Thrown when reading a package manifest fails

Fields of PackageReadError

file: PathBuf

PackageUnpackError

Thrown when a package has been unpacked but is not formed correctly.

PackageWriteError

Thrown when writing a package manifest fails

Fields of PackageWriteError

file: PathBuf

ParseBinConfigError

Thrown when unable to parse a bin config file

ParseHooksError

Thrown when unable to parse a hooks.json file

Fields of ParseHooksError

file: PathBuf

ParseNodeIndexCacheError

Thrown when unable to parse the node index cache

ParseNodeIndexError

Thrown when unable to parse the node index

Fields of ParseNodeIndexError

from_url: String

ParseNodeIndexExpiryError

Thrown when unable to parse the node index cache expiration

ParseNpmManifestError

Thrown when unable to parse the npm manifest file from a node install

ParsePackageConfigError

Thrown when unable to parse a package configuration

ParsePlatformError

Thrown when unable to parse the platform.json file

ParseToolSpecError

Thrown when unable to parse a tool spec (<tool>[@<version>])

Fields of ParseToolSpecError

tool_spec: String

PersistInventoryError

Thrown when persisting an archive to the inventory fails

Fields of PersistInventoryError

tool: String

ProjectLocalBinaryExecError

Thrown when executing a project-local binary fails

Fields of ProjectLocalBinaryExecError

command: String

ProjectLocalBinaryNotFound

Thrown when a project-local binary could not be found

Fields of ProjectLocalBinaryNotFound

command: String

PublishHookBothUrlAndBin

Thrown when a publish hook contains both the url and bin fields

PublishHookNeitherUrlNorBin

Thrown when a publish hook contains neither url nor bin fields

ReadBinConfigDirError

Thrown when there was an error reading the user bin directory

Fields of ReadBinConfigDirError

dir: PathBuf

ReadBinConfigError

Thrown when there was an error reading the config for a binary

Fields of ReadBinConfigError

file: PathBuf

ReadDefaultNpmError

Thrown when unable to read the default npm version file

Fields of ReadDefaultNpmError

file: PathBuf

ReadDirError

Thrown when unable to read the contents of a directory

Fields of ReadDirError

dir: PathBuf

ReadHooksError

Thrown when there was an error opening a hooks.json file

Fields of ReadHooksError

file: PathBuf

ReadNodeIndexCacheError

Thrown when there was an error reading the Node Index Cache

Fields of ReadNodeIndexCacheError

file: PathBuf

ReadNodeIndexExpiryError

Thrown when there was an error reading the Node Index Cache Expiration

Fields of ReadNodeIndexExpiryError

file: PathBuf

ReadNpmManifestError

Thrown when there was an error reading the npm manifest file

ReadPackageConfigError

Thrown when there was an error reading a package configuration file

Fields of ReadPackageConfigError

file: PathBuf

ReadPlatformError

Thrown when there was an error opening the user platform file

Fields of ReadPlatformError

file: PathBuf

RegistryFetchError

Thrown when the public registry for Node or Yarn could not be downloaded.

Fields of RegistryFetchError

tool: Stringfrom_url: String

RunShimDirectly

Thrown when the shim binary is called directly, not through a symlink

SetupToolImageError

Thrown when there was an error copying an unpacked tool to the image directory

Fields of SetupToolImageError

tool: Stringversion: Stringdir: PathBuf

ShimCreateError

Thrown when Volta is unable to create a shim

Fields of ShimCreateError

name: String

ShimRemoveError

Thrown when Volta is unable to remove a shim

Fields of ShimRemoveError

name: String

StringifyBinConfigError

Thrown when serializnig a bin config to JSON fails

StringifyPackageConfigError

Thrown when serializnig a package config to JSON fails

StringifyPlatformError

Thrown when serializing the platform to JSON fails

Unimplemented

Thrown when a given feature has not yet been implemented

Fields of Unimplemented

feature: String

UnpackArchiveError

Thrown when unpacking an archive (tarball or zip) fails

Fields of UnpackArchiveError

tool: Stringversion: String

UpgradePackageNotFound

Thrown when a package to upgrade was not found

Fields of UpgradePackageNotFound

package: Stringmanager: PackageManager

UpgradePackageWrongManager

Thrown when a package to upgrade was installed with a different package manager

Fields of UpgradePackageWrongManager

package: Stringmanager: PackageManager

VersionParseError

Fields of VersionParseError

version: String

WriteBinConfigError

Thrown when there was an error writing a bin config file

Fields of WriteBinConfigError

file: PathBuf

WriteDefaultNpmError

Thrown when there was an error writing the default npm to file

Fields of WriteDefaultNpmError

file: PathBuf

WriteLauncherError

Thrown when there was an error writing the npm launcher

Fields of WriteLauncherError

tool: String

WriteNodeIndexCacheError

Thrown when there was an error writing the node index cache

Fields of WriteNodeIndexCacheError

file: PathBuf

WriteNodeIndexExpiryError

Thrown when there was an error writing the node index expiration

Fields of WriteNodeIndexExpiryError

file: PathBuf

WritePackageConfigError

Thrown when there was an error writing a package config

Fields of WritePackageConfigError

file: PathBuf

WritePlatformError

Thrown when writing the platform.json file fails

Fields of WritePlatformError

file: PathBuf

YarnLatestFetchError

Thrown when there is an error fetching the latest version of Yarn

Fields of YarnLatestFetchError

from_url: String

YarnVersionNotFound

Thrown when there is no Yarn version matching a requested semver specifier.

Fields of YarnVersionNotFound

matching: String

Implementations

impl ErrorKind

pub fn exit_code(&self) -> ExitCode