volta_core::fs

Function read_dir_eager

pub fn read_dir_eager(
    dir: &Path,
) -> Result<impl Iterator<Item = (DirEntry, Metadata)>>

Reads the full contents of a directory, eagerly extracting each directory entry and its metadata and returning an iterator over them. Returns Error if any of these steps fails.

This function makes it easier to write high level logic for manipulating the contents of directories (map, filter, etc).

Note that this function allocates an intermediate vector of directory entries to construct the iterator from, so if a directory is expected to be very large, it will allocate temporary data proportional to the number of entries.