Module volta_core::log

This module provides a custom Logger implementation for use with the log crate

Structs

Logger

Enums

LogContext
Represents the context from which the logger was created

LogVerbosity
Represents the level of verbosity that was requested by the user

Constants

ALLOWED_PREFIX

ERROR_PREFIX

MIGRATION_ERROR_PREFIX

MIGRATION_WARNING_PREFIX

SHIM_ERROR_PREFIX

SHIM_WARNING_PREFIX

VOLTA_LOGLEVEL

WARNING_PREFIX

WRAP_INDENT

Functions

level_from_env
Determines the correct logging level based on the environment If VOLTA_LOGLEVEL is set to a valid level, we use that If not, we check the current stdout to determine whether it is a TTY or not If it is a TTY, we use Info If it is NOT a TTY, we use Error as we don't want to show warnings when running as a script

wrap_content
Wraps the supplied content to the terminal width, if we are in a terminal. If not, returns the content as a String