Module volta_core::path

Provides functions for determining the paths of files and directories in a standard Volta layout.

Modules

unix	Unix
Provides functions for determining the paths of files and directories in a standard Volta layout in Unix-based operating systems.

windows	Windows
Provides functions for determining the paths of files and directories in a standard Volta layout in Windows operating systems.

Constants

ARCH
The system architecture component of a Node distribution tarball's name.

OS
The OS component of a Node distribution tarball's name.

Functions

archive_extension

cache_dir

create_file_symlink
Create a symlink. The dst path will be a symbolic link pointing to the src path.

default_volta_home

ensure_dir_exists

ensure_shim_exists

ensure_volta_dirs_exist

env_paths

find_project_dir

image_dir

inventory_dir

is_dependency

is_node_modules

is_node_root

is_project_root

log_dir

node_archive_npm_package_json_path

node_archive_root_dir_name

node_cache_dir

node_distro_file_name

node_image_bin_dir

node_image_dir

node_image_root_dir

node_index_expiry_file

node_index_file

node_inventory_dir

node_npm_version_file

package_archive_root_dir_name

package_distro_file

package_distro_file_name

package_distro_shasum

package_image_dir

package_image_root_dir

package_inventory_dir

package_shasum_file_name

shim_dir

shim_executable

shim_file

tmp_dir

tools_dir

user_bin_dir

user_hooks_file

user_package_config_file

user_package_dir

user_platform_file

user_tool_bin_config

user_toolchain_dir

volta_file

volta_home

yarn_archive_root_dir_name

yarn_distro_file_name

yarn_image_bin_dir

yarn_image_dir

yarn_image_root_dir

yarn_inventory_dir