volta_core

Module run

Modules

binary

executor 🔒

node 🔒

npm 🔒

npx 🔒

parser 🔒

pnpm 🔒

yarn 🔒

Constants

RECURSION_ENV_VAR 🔒

Environment variable set internally when a shim has been executed and the context evaluated

VOLTA_BYPASS 🔒

Functions

debug_active_image 🔒

Write a debug message with the full image that will be used to execute a command

debug_no_platform 🔒

Write a debug message that there is no platform available

execute_shim

Execute a shim command, based on the command-line arguments to the current process

execute_tool

Execute a tool with the provided arguments

format_tool_version 🔒

get_executor 🔒

Get the appropriate Tool command, based on the requested executable and arguments

get_tool_name 🔒

Determine the name of the command to run by inspecting the first argument to the active process

tool_name_from_file_name 🔒