Struct volta_core::run::executor::PackageUpgradeCommand

pub struct PackageUpgradeCommand {
    command: Command,
    upgrader: InPlaceUpgrade,
    platform: Platform,
}

Process builder for launching a global package upgrade command (e.g. npm update -g)

This will use an InPlaceUpgrade instance to modify the command and point at the appropriate image directory. It will also complete the install, writing any updated configs and shims

Fields
command: Command

The command that will ultimately be executed

upgrader: InPlaceUpgrade

Helper utility to modify the command and provide the completion method

platform: Platform

The platform to run the command under

Implementations

impl PackageUpgradeCommand

pub fn new<A, S>( args: A, package: String, platform: Platform, manager: PackageManager ) -> Fallible<Self>where A: IntoIterator<Item = S>, S: AsRef<OsStr>,

pub fn envs<E, K, V>(&mut self, envs: E)where E: IntoIterator<Item = (K, V)>, K: AsRef<OsStr>, V: AsRef<OsStr>,

Adds or updates environment variables that the command will use

pub fn cli_platform(&mut self, cli: CliPlatform)

Updates the Platform for the command to include values from the command-line

pub fn execute(self, session: &mut Session) -> Fallible<ExitStatus>

Runs the upgrade command, applying the necessary modifications to point at the Volta image directory

Will also check for common failure cases, such as non-existant package or wrong package manager