volta_core::run

Module parser

Structs

InstallArgs

The arguments passed to a global install command

LinkArgs

The arguments passed to an npm link command

UninstallArgs

The list of tools passed to an uninstall command

UpgradeArgs

The list of tools passed to an upgrade command

Enums

CommandArg

GlobalCommand

InterceptedCommand

An intercepted local command

Constants

NPM_INSTALL_ALIASES 🔒

Aliases that npm supports for the ‘install’ command

NPM_LINK_ALIASES 🔒

Aliases that npm supports for the ‘link’ command

NPM_UNINSTALL_ALIASES 🔒

Aliases that npm supports for the ‘uninstall’ command

NPM_UPDATE_ALIASES 🔒

Aliases that npm supports for the update command

PNPM_LINK_ALIASES 🔒

Aliases that pnpm supports for the ‘link’ command see: https://pnpm.io/cli/link

PNPM_UNINSTALL_ALIASES 🔒

Aliases that pnpm supports for the ‘remove’ command, see: https://pnpm.io/cli/remove

PNPM_UPDATE_ALIASES 🔒

Aliases that pnpm supports for the ‘update’ command, see: https://pnpm.io/cli/update

UNSAFE_GLOBAL 🔒

Functions

has_global_without_prefix 🔒

Check if the provided argument list includes a global flag and doesn’t have a prefix setting

is_flag 🔒

is_positional 🔒