volta_core::run::parser

Function has_global_without_prefix

fn has_global_without_prefix<A>(args: &[A]) -> bool
where
    A: AsRef<OsStr>,

Check if the provided argument list includes a global flag and doesn’t have a prefix setting

For our interception, we only want to intercept global commands. Additionally, if the user passes a prefix setting, that will override the logic we use to redirect the install, so our process won’t work and will cause an error. We should avoid intercepting in those cases since a command with an explicit prefix is something beyond the “standard” global install anyway.