Struct volta_core::run::ToolCommand

pub(crate) struct ToolCommand {
    command: Command,
    on_failure: ErrorDetails,
}

Represents the command to execute a tool

Fields
command: Command

The command that will execute a tool with the right PATH context

on_failure: ErrorDetails

The Volta error with which to wrap any failure.

This allows us to call out to the system for the pass-through behavior, but still show a friendly error message for cases where the user needs to select a Node version

Methods

impl ToolCommand

fn direct<A>(exe: &OsStr, args: A, path_var: &OsStr) -> Self where
A: IntoIterator<Item = OsString>,

Build a ToolCommand that is directly calling a tool in the Volta directory

fn project_local<A>(exe: &OsStr, args: A, path_var: &OsStr) -> Self where
A: IntoIterator<Item = OsString>,

Build a ToolCommand that is calling a binary in the current project's node_modules/bin

fn passthrough<A>(
exe: &OsStr,
args: A,
default_error: ErrorDetails
) -> Fallible<Self> where
A: IntoIterator<Item = OsString>,

Build a ToolCommand that is calling a command that Volta couldn't find

This will allow the existing system to resolve the tool, if possible. If that still fails, then we show default_error as the friendly error to the user, directing them how to resolve the issue (e.g. run volta install node to enable node)

pub(crate) fn current_dir<P: AsRef<Path>>(&mut self, dir: P) -> &mut ToolCommand

pub(crate) fn status(self) -> Fallible<ExitStatus>

pub(crate) fn output(self) -> Fallible<Output>