volta_core::session

Struct Session

pub struct Session {
    hooks: LazyHookConfig,
    toolchain: LazyToolchain,
    project: LazyProject,
    event_log: EventLog,
}

Represents the user’s state during an execution of a Volta tool. The session encapsulates a number of aspects of the environment in which the tool was invoked, including:

the current directory

the Node project tree that contains the current directory (if any)

the Volta hook configuration

the inventory of locally-fetched Volta tools

Fields
hooks: LazyHookConfigtoolchain: LazyToolchainproject: LazyProjectevent_log: EventLog
Implementations

impl Session

pub fn init() -> Session

Constructs a new Session.

pub fn project(&self) -> Fallible<Option<&Project>>

Produces a reference to the current Node project, if any.

pub fn project_mut(&mut self) -> Fallible<Option<&mut Project>>

Produces a mutable reference to the current Node project, if any.

pub fn default_platform(&self) -> Fallible<Option<&PlatformSpec>>

Returns the user’s default platform, if any

pub fn project_platform(&self) -> Fallible<Option<&PlatformSpec>>

Returns the current project’s pinned platform image, if any.

pub fn toolchain(&self) -> Fallible<&Toolchain>

Produces a reference to the current toolchain (default platform specification)

pub fn toolchain_mut(&mut self) -> Fallible<&mut Toolchain>

Produces a mutable reference to the current toolchain

pub fn hooks(&self) -> Fallible<&HookConfig>

Produces a reference to the hook configuration

pub fn add_event_start(&mut self, activity_kind: ActivityKind)

pub fn add_event_end( &mut self, activity_kind: ActivityKind, exit_code: ExitCode, )

pub fn add_event_tool_end( &mut self, activity_kind: ActivityKind, exit_code: i32, )

pub fn add_event_error( &mut self, activity_kind: ActivityKind, error: &VoltaError, )

fn publish_to_event_log(self)

pub fn exit(self, code: ExitCode) -> !

pub fn exit_tool(self, code: i32) -> !