volta_core::style

Function progress_bar

pub fn progress_bar(origin: Origin, details: &str, len: u64) -> ProgressBar

Constructs a command-line progress bar based on the specified Origin enum (e.g., Origin::Remote), details string (e.g., "v1.23.4"), and logical length (i.e., the number of logical progress steps in the process being visualized by the progress bar).