Module volta_core::sync

Inter-process locking on the Volta directory

To avoid issues where multiple separate invocations of Volta modify the data directory simultaneously, we provide a locking mechanism that only allows a single process to modify the directory at a time.

However, within a single process, we may attempt to lock the directory in different code paths. For example, when installing a package we require a lock, however we also may need to install Node, which requires a lock as well. To avoid deadlocks in those situations, we track the state of the lock globally:

If a lock is requested and no locks are active, then we acquire a file lock on the volta.lock file and initialize the state with a count of 1

If a lock already exists, then we increment the count of active locks

When a lock is no longer needed, we decrement the count of active locks

When the last lock is released, we release the file lock and clear the global lock state.

This allows multiple code paths to request a lock and not worry about potential deadlocks, while still preventing multiple processes from making concurrent changes.

Structs

LOCK_STATE

LockState
The current state of locks for this process.

VoltaLock
An RAII implementation of a process lock on the Volta directory. A given Volta process can have multiple active locks, but only one process can have any locks at a time.

Constants

LOCK_FILE