volta_core

Module tool

Re-exports

pub use node::load_default_npm_version;

pub use node::Node;

pub use node::NODE_DISTRO_ARCH;

pub use node::NODE_DISTRO_EXTENSION;

pub use node::NODE_DISTRO_OS;

pub use npm::BundledNpm;

pub use npm::Npm;

pub use package::BinConfig;

pub use package::Package;

pub use package::PackageConfig;

pub use package::PackageManifest;

pub use pnpm::Pnpm;

pub use yarn::Yarn;

Modules

node

npm

package

pnpm

registry 🔒

serial 🔒

yarn

Structs

PackageDetails

Details about a package in the npm Registry

Enums

FetchStatus 🔒

Represents the result of checking if a tool is available locally or not

Spec

Specification for a tool and its associated version.

Constants

PATH_VAR_NAME 🔒

Traits

Tool

Trait representing all of the actions that can be taken with a tool

Functions

check_fetched 🔒

Uses the supplied already_fetched predicate to determine if a tool is available or not.

check_shim_reachable

Check if a newly-installed shim is first on the PATH. If it isn’t, we want to inform the user that they’ll want to move it to the start of PATH to make sure things work as expected.

debug_already_fetched 🔒

download_tool_error 🔒

find_expected_shim_dir 🔒

Locate the base directory for the relevant shim in the Volta directories.

info_fetched 🔒

info_installed 🔒

info_pinned 🔒

info_project_version 🔒

registry_fetch_error 🔒