Module volta_core::tool::node::resolve

Provides resolution of Node requirements into specific versions, using the NodeJS index

Functions

match_node_version 🔒

max_age 🔒

Get the cache max-age of an HTTP reponse.

public_node_version_index 🔒

Returns the URL of the index of available Node versions on the public Node server.

read_cached_opt 🔒

Reads a public index from the Node cache, if it exists and hasn’t expired.

resolve

resolve_latest 🔒

resolve_lts 🔒

resolve_node_versions 🔒

resolve_semver 🔒