volta_core::tool

Module package

Modules

configure 🔒

install 🔒

manager 🔒

metadata 🔒

uninstall 🔒

Structs

BinConfig

Configuration information about a single installed binary from a package

DirectInstall

Helper struct for direct installs through npm i -g or yarn global add

InPlaceUpgrade

Helper struct for direct in-place upgrades using npm update -g or yarn global upgrade

Package

The Tool implementation for installing 3rd-party global packages

PackageConfig

Configuration information about an installed package

PackageManifest

The relevant information we need out of a package’s package.json file

Enums

NeedsScope 🔒

PackageManager

The package manager used to install a given package

Functions

link_package_to_shared_dir 🔒

persist_install 🔒

setup_staging_directory 🔒

Create the temporary staging directory we will use to install and ensure expected subdirectories exist within it

uninstall

Uninstalls the specified package.