Module volta_core::tool::package::install

Provides installation steps for 3rd-party packages, fetching their dependencies, writing config files, and creating shims

Modules

os

Structs

BinConfig
Configuration information about an installed binary from a package.

BinLoader
Information about the Shebang script loader (e.g. #!/usr/bin/env node)

PackageConfig
Configuration information about an installed package.

SHEBANG

Functions

build_install_command

determine_engine

determine_script_loader
Read the script for a shebang loader. If found, return it so it will be stored in the config

install

install_dependencies

read_bins
Read a fetched package and generate a map of all the bins it provides

write_configs