volta_core::tool::package

Module manager

Enums

PackageManager

The package manager used to install a given package

Functions

get_npm_package_name 🔒

Determine the package name for an npm global install

get_pnpm_or_yarn_package_name 🔒

Determine the package name for a pnpm or Yarn global install

get_single_directory_name 🔒

Return the name of the single subdirectory (if any) to the given parent_dir