Struct volta_core::tool::package::DirectInstall

pub struct DirectInstall {
    staging: TempDir,
    manager: PackageManager,
    name: Option<String>,
}

Helper struct for direct installs through npm i -g or yarn global add

Provides methods to simplify installing into a staging directory and then moving that install into the proper location after it is complete.

Note: We don’t always know the name of the package up-front, as the install could be from a tarball or a git coordinate. If we do know ahead of time, then we can skip looking it up

Fields
staging: TempDirmanager: PackageManagername: Option<String>
Implementations

impl DirectInstall

pub fn new(manager: PackageManager) -> Fallible<Self>

pub fn with_name(manager: PackageManager, name: String) -> Fallible<Self>

pub fn setup_command(&self, command: &mut Command)

pub fn complete_install(self, image: &Image) -> Fallible<()>