volta_core::tool::package

Struct InPlaceUpgrade

pub struct InPlaceUpgrade {
    package: String,
    directory: PathBuf,
    manager: PackageManager,
}

Helper struct for direct in-place upgrades using npm update -g or yarn global upgrade

Upgrades the requested package directly in the image directory

Fields
package: Stringdirectory: PathBufmanager: PackageManager
Implementations

impl InPlaceUpgrade

pub fn new(package: String, manager: PackageManager) -> Fallible<Self>

pub fn check_upgraded_package(&self) -> Fallible<()>

Check for possible failure cases with the package to be upgraded - The package is not installed as a global - The package exists, but was installed with a different package manager

pub fn setup_command(&self, command: &mut Command)

pub fn complete_upgrade(self, image: &Image) -> Fallible<()>