Crate volta_core

The main implementation crate for the core of Volta.

Modules

command 🔒

error

event

Events for the sessions in executables and shims and everything

fs

Provides utilities for operating on the filesystem.

hook 🔒

Provides types for working with Volta hooks.

inventory

Provides types for working with Volta’s inventory, the local repository of available tool versions.

layout

log

This module provides a custom Logger implementation for use with the log crate

monitor

platform

project

Provides the Project type, which represents a Node project tree in the filesystem.

run

session

Provides the Session type, which represents the user’s state during an execution of a Volta tool, including their current directory, Volta hook configuration, and the state of the local inventory.

shim

Provides utilities for modifying shims for 3rd-party executables

signal

style

The view layer of Volta, with utilities for styling command-line output.

sync

Inter-process locking on the Volta directory

tool

toolchain

version

Constants

VOLTA_FEATURE_PNPM 🔒