Module volta_core::fs

Provides utilities for operating on the filesystem.

Functions

create_staging_dir
Creates a staging directory in the Volta tmp directory

create_staging_file
Creates a NamedTempFile in the Volta tmp directory

delete_dir_error

delete_file_error

dir_entry_match
Reads the contents of a directory and returns a Vec of the matched results from the input function

ensure_dir_does_not_exist
This deletes the input directory, if it exists

read_dir_eager
Reads the full contents of a directory, eagerly extracting each directory entry and its metadata and returning an iterator over them. Returns Error if any of these steps fails.

read_file
Reads a file, if it exists.

symlink_file
Create a symlink. The dst path will be a symbolic link pointing to the src path.

touch
Opens a file, creating it if it doesn't exist