Module volta_core::hook::tool

Types representing Volta Tool Hooks.

Structs

REL_PATH 🔒

REL_PATH_PARENT 🔒

YarnIndexHook

A hook for resolving the URL for the Yarn index

Enums

DistroHook

A hook for resolving the distro URL for a given tool version

MetadataHook

A hook for resolving the URL for metadata about a tool

Constants

ARCH_TEMPLATE 🔒

EXTENSION_TEMPLATE 🔒

FILENAME_TEMPLATE 🔒

OS_TEMPLATE 🔒

VERSION_TEMPLATE 🔒

Functions

calculate_extension 🔒

Use the expected filename to determine the extension for this hook

execute_binary 🔒

Execute a shell command and return the trimmed stdout from that command