Module volta_core::inventory

Provides types for working with Volta’s inventory, the local repository of available tool versions.

Functions

node_available

Checks if a given Node version image is available on the local machine

node_versions

Collects a set of all Node versions fetched on the local machine

npm_available

Checks if a given npm version image is available on the local machine

npm_versions

Collects a set of all npm versions fetched on the local machine

package_configs

Collects a set of all Package Configs on the local machine

read_versions

Reads the contents of a directory and returns the set of all versions found in the directory’s listing by parsing the directory names as semantic versions

yarn_available

Checks if a given Yarn version image is available on the local machine

yarn_versions

Collects a set of all Yarn versions fetched on the local machine