Module volta_core::platform

Modules

image

system

Structs

BinaryPlatformSpec

DefaultPlatformSpec

Image
A platform image.

MergedPlatformSpec

ProjectPlatformSpec

Sourced

System
A lightweight namespace type representing the system environment, i.e. the environment with Volta removed.

Enums

Source
The source with which a version is associated

Traits

PlatformSpec

Functions

build_path_error

ensure_node
Ensures that a specific Node version has been fetched and unpacked

ensure_yarn
Ensures that a specific Yarn version has been fetched and unpacked