Module volta_core::platform

Modules

image

system

Structs

CliPlatform
Represents a (maybe) platform with values from the command line

Image
A platform image.

Platform
Represents a real Platform, with Versions pulled from one or more PlatformSpecs

PlatformSpec
Represents the specification of a single Platform, regardless of the source

Sourced

System
A lightweight namespace type representing the system environment, i.e. the environment with Volta removed.

Enums

InheritOption
Represents 3 possible states: Having a value, not having a value, and inheriting a value

Source
The source with which a version is associated

Functions

build_path_error