Function volta_core::project::serial::update_manifest

pub(super) fn update_manifest(
    file: &Path,
    key: ManifestKey,
    value: Option<&Version>
) -> Fallible<()>

Updates the volta hash in the specified manifest with the given key and value

Will create the volta hash if it isn’t already present

If the value is None, will remove the key from the hash