Struct volta_core::project::Project

pub struct Project {
    manifest_file: PathBuf,
    workspace_manifests: IndexSet<PathBuf>,
    dependencies: ChainMap<String, String>,
    platform: Option<PlatformSpec>,
}

A Node project workspace in the filesystem

Fields
manifest_file: PathBufworkspace_manifests: IndexSet<PathBuf>dependencies: ChainMap<String, String>platform: Option<PlatformSpec>
Implementations

impl Project

fn for_current_dir() -> Fallible<Option<Self>>

Creates an optional Project instance from the current directory

fn for_dir(base_dir: PathBuf) -> Fallible<Option<Self>>

Creates an optional Project instance from the specified directory

Will search ancestors to find a package.json and use that as the root of the project

fn from_file(manifest_file: PathBuf) -> Fallible<Self>

Creates a Project instance from the given package manifest file (package.json)

pub fn manifest_file(&self) -> &Path

Returns a reference to the manifest file for the current project

pub fn workspace_roots(&self) -> impl Iterator<Item = &Path>

Returns an iterator of paths to all of the workspace roots

pub fn platform(&self) -> Option<&PlatformSpec>

Returns a reference to the Project's PlatformSpec, if available

pub fn has_direct_dependency(&self, dependency: &str) -> bool

Returns true if the project dependency map contains the specified dependency

pub fn has_direct_bin(&self, bin_name: &OsStr) -> Fallible<bool>

Returns true if the input binary name is a direct dependency of the input project

pub fn find_bin<P: AsRef<Path>>(&self, bin_name: P) -> Option<PathBuf>

Searches the project roots to find the path to a project-local binary file

pub fn pin_node(&mut self, version: Version) -> Fallible<()>

Pins the Node version in this project's manifest file

pub fn pin_npm(&mut self, version: Option<Version>) -> Fallible<()>

Pins the npm version in this project's manifest file

pub fn pin_yarn(&mut self, version: Option<Version>) -> Fallible<()>

Pins the Yarn version in this project's manifest file