Struct volta_core::project::Project

pub struct Project {
    manifest: Manifest,
    project_root: PathBuf,
}

A Node project tree in the filesystem.

Fields
manifest: Manifestproject_root: PathBuf
Methods

impl Project

fn for_current_dir() -> Fallible<Option<Project>>

Returns the Node project containing the current working directory, if any.

pub(crate) fn find_dir(base_dir: &Path) -> Option<&Path>

Starts at base_dir and walks up the directory tree until a package.json file is found

fn for_dir(base_dir: &Path) -> Fallible<Option<Project>>

Returns the Node project for the input directory, if any.

pub fn platform(&self) -> Option<Rc<PlatformSpec>>

Returns the pinned platform image, if any.

pub fn is_pinned(&self) -> bool

Returns true if the project manifest contains a toolchain.

pub fn manifest(&self) -> &Manifest

Returns the project manifest (package.json) for this project.

pub fn package_file(&self) -> PathBuf

Returns the path to the package.json file for this project.

pub fn local_bin_dir(&self) -> PathBuf

Returns the path to the local binary directory for this project.

pub fn has_direct_bin(&self, bin_name: &OsStr) -> Fallible<bool>

Returns true if the input binary name is a direct dependency of the input project

pub fn matching_bin(
&self,
bin_name: &OsStr,
version: &Version
) -> Fallible<Option<BinConfig>>

Returns a matching config if the bin exists at the specified version in the project.

fn has_direct_dependency(&self, dependency: &str) -> bool

pub fn has_dependency(&self, dependency: &str, version: &Version) -> bool

pub fn pin_node(&mut self, node_version: &NodeVersion) -> Fallible<()>

Writes the specified version of Node to the volta.node key in package.json.

pub fn pin_yarn(&mut self, yarn_version: &Version) -> Fallible<()>

Writes the specified version of Yarn to the volta.yarn key in package.json.

pub fn pin_npm(&mut self, npm_version: &Version) -> Fallible<()>

Writes the specified version of Npm to the volta.npm key in package.json.