Module volta_core::run::executor

Structs

InternalInstallCommand

Executor for running an internal install (installing Node, npm, or Yarn using the volta install logic)

PackageInstallCommand

Process builder for launching a package install command (e.g. npm install --global)

PackageLinkCommand

Process builder for launching a npm link <package> command

PackageUpgradeCommand

Process builder for launching a global package upgrade command (e.g. npm update -g)

ToolCommand

Process builder for launching a Volta-managed tool

UninstallCommand

Executor for running a tool uninstall command.

Enums

Executor

ToolKind

The kind of tool being executed, used to determine the correct execution context