Module volta_core::run::parser

Structs

InstallArgs

The arguments passed to a global install command

LinkArgs

The arguments passed to an npm link command

UninstallArgs

The list of tools passed to an uninstall command

UpgradeArgs

The list of tools passed to an upgrade command

Enums

CommandArg

GlobalCommand

InterceptedCommand

An intercepted local command

Constants

NPM_INSTALL_ALIASES

Aliases that npm supports for the ‘install’ command

NPM_LINK_ALIASES

Aliases that npm supports for the ‘link’ command

NPM_UNINSTALL_ALIASES

Aliases that npm supports for the ‘uninstall’ command

NPM_UPDATE_ALIASES

Aliases that npm supports for the update command

UNSAFE_GLOBAL

Functions

has_global_flag

is_flag

is_positional