Struct volta_core::session::Session

pub struct Session {
    hooks: LazyHookConfig,
    inventory: LazyInventory,
    toolchain: LazyToolchain,
    project: LazyProject,
    event_log: EventLog,
}

Represents the user's state during an execution of a Volta tool. The session encapsulates a number of aspects of the environment in which the tool was invoked, including:

the current directory

the Node project tree that contains the current directory (if any)

the Volta hook configuration

the inventory of locally-fetched Volta tools

Fields
hooks: LazyHookConfiginventory: LazyInventorytoolchain: LazyToolchainproject: LazyProjectevent_log: EventLog
Methods

impl Session

pub fn init() -> Session

Constructs a new Session.

pub fn project(&self) -> Fallible<Option<&Project>>

Produces a reference to the current Node project, if any.

pub fn project_mut(&mut self) -> Fallible<Option<&mut Project>>

Produces a mutable reference to the current Node project, if any.

pub fn current_platform(&self) -> Fallible<Option<SourcedPlatformSpec>>

Returns the user's currently active platform, if any

Active platform is determined by first looking at the Project Platform

If it exists and has a Yarn version, then we use the project platform

If it exists but doesn't have a Yarn version, then we merge the two, pulling Yarn from the user default platform, if available

If there is no Project platform, then we use the user Default Platform

pub fn default_platform(&self) -> Fallible<Option<Rc<PlatformSpec>>>

Returns the user's default platform, if any

pub fn project_platform(&self) -> Fallible<Option<Rc<PlatformSpec>>>

Returns the current project's pinned platform image, if any.

pub fn inventory(&self) -> Fallible<&Inventory>

Produces a reference to the current inventory.

pub fn inventory_mut(&mut self) -> Fallible<&mut Inventory>

Produces a mutable reference to the current inventory.

pub fn toolchain(&self) -> Fallible<&Toolchain>

Produces a reference to the current toolchain (default platform specification)

pub fn toolchain_mut(&mut self) -> Fallible<&mut Toolchain>

Produces a mutable reference to the current toolchain

pub fn hooks(&self) -> Fallible<&HookConfig>

Produces a reference to the hook configuration

pub(crate) fn ensure_node(&mut self, version: &Version) -> Fallible<()>

Ensures that a specific Node version has been fetched and unpacked

pub(crate) fn ensure_yarn(&mut self, version: &Version) -> Fallible<()>

Ensures that a specific Yarn version has been fetched and unpacked

pub fn add_event_start(&mut self, activity_kind: ActivityKind)

pub fn add_event_end(
&mut self,
activity_kind: ActivityKind,
exit_code: ExitCode
)

pub fn add_event_tool_end(
&mut self,
activity_kind: ActivityKind,
exit_code: i32
)

pub fn add_event_error(
&mut self,
activity_kind: ActivityKind,
error: &VoltaError
)

fn publish_to_event_log(self)

pub fn exit(self, code: ExitCode) -> !

pub fn exit_tool(self, code: i32) -> !