Module volta_core::style

The view layer of Volta, with utilities for styling command-line output.

Constants

MAX_PROGRESS_WIDTH

MAX_WIDTH

Functions

action_str

Determines the string to display based on the Origin of the operation.

format_error_cause

Format the underlying cause of an error

note_prefix

Generate the styled prefix for a note

progress_bar

Constructs a command-line progress bar based on the specified Origin enum (e.g., Origin::Remote), details string (e.g., "v1.23.4"), and logical length (i.e., the number of logical progress steps in the process being visualized by the progress bar).

progress_spinner

Constructs a command-line progress spinner with the specified “message” string. The spinner is ticked by default every 50ms.

success_prefix

Generate the styled prefix for a success message

text_width

Get the width of the terminal, limited to a maximum of MAX_WIDTH

tool_version