Module volta_core::tool

Re-exports

pub use node::Node;

pub use node::NODE_DISTRO_ARCH;

pub use node::NODE_DISTRO_EXTENSION;

pub use node::NODE_DISTRO_OS;

pub use npm::BundledNpm;

pub use npm::Npm;

pub use package::bin_full_path;

pub use package::Package;

pub use package::PackageDetails;

pub use yarn::Yarn;

Modules

node

npm

package

serial

yarn

Structs

BinConfig
Configuration information about an installed binary from a package.

BinLoader
Information about the Shebang script loader (e.g. #!/usr/bin/env node)

PackageConfig
Configuration information about an installed package.

Enums

Spec
Specification for a tool and its associated version.

Traits

Tool
Trait representing all of the actions that can be taken with a tool

Functions

debug_already_fetched

download_tool_error

info_fetched

info_installed

info_pinned

info_project_version

load_default_npm_version
Load the local npm version file to determine the default npm version for a given version of Node

registry_fetch_error