Module volta_core::tool

Re-exports

pub use node::Node;

pub use node::NODE_DISTRO_ARCH;

pub use node::NODE_DISTRO_EXTENSION;

pub use node::NODE_DISTRO_OS;

pub use npm::BundledNpm;

pub use npm::Npm;

pub use package::Package;

pub use yarn::Yarn;

Modules

node

npm

package

registry 🔒

serial 🔒

yarn

Structs

BinConfig

Configuration information about a single installed binary from a package

PackageConfig

Configuration information about an installed package

PackageDetails

Details about a package in the npm Registry

PackageManifest

The relevant information we need out of a package’s package.json file

Enums

FetchStatus 🔒

Represents the result of checking if a tool is available locally or not

Spec

Specification for a tool and its associated version.

Traits

Tool

Trait representing all of the actions that can be taken with a tool

Functions

check_fetched 🔒

Uses the supplied already_fetched predicate to determine if a tool is available or not.

debug_already_fetched 🔒

download_tool_error 🔒

info_fetched 🔒

info_installed 🔒

info_pinned 🔒

info_project_version 🔒

load_default_npm_version

Load the local npm version file to determine the default npm version for a given version of Node

registry_fetch_error 🔒