Module volta_core::tool

Modules

node

npm

package

serial

yarn

Structs

BinConfig
Configuration information about an installed binary from a package.

BinLoader
Information about the Shebang script loader (e.g. #!/usr/bin/env node)

Node
The Tool implementation for fetching and installing Node

NodeVersion
A full Node version including not just the version of Node itself but also the specific version of npm installed globally with that Node installation.

Npm
The Tool implementation for fetching and installing Npm (Unimplemented)

Package
The Tool implementation for fetching and installing 3rd-party packages

PackageConfig
Configuration information about an installed package.

PackageDetails
Details required for fetching a 3rd-party Package

Yarn
The Tool implementation for fetching and installing Yarn

Enums

Resolved
A fully resolved Tool, with all information necessary for fetching

Spec
Specification for a tool and its associated version.

Constants

NODE_DISTRO_ARCH
The system architecture component of a Node distro's name.

NODE_DISTRO_EXTENSION
Filename extension for Node distro files.

NODE_DISTRO_OS
The OS component of a Node distro's filename.

Traits

Tool
Trait representing all of the actions that can be taken with a tool

Functions

bin_full_path

debug_already_fetched

download_tool_error

info_fetched

info_installed

info_pinned

info_project_version

load_default_npm_version
Load the local npm version file to determine the default npm version for a given version of Node

registry_fetch_error