Enum volta_core::tool::Spec

pub enum Spec {
    Node(VersionSpec),
    Npm(VersionSpec),
    Yarn(VersionSpec),
    Package(String, VersionSpec),
}

Specification for a tool and its associated version.

Variants

Node(VersionSpec)

Npm(VersionSpec)

Yarn(VersionSpec)

Package(String, VersionSpec)

Methods

impl Spec

Methods for parsing a Spec out of string values

pub fn from_str_and_version(tool_name: &str, version: VersionSpec) -> Self

pub fn try_from_str(tool_spec: &str) -> Fallible<Self>

Try to parse a tool and version from a string like `[@].

pub fn from_strings<T>(tool_strs: &[T], action: &str) -> Fallible<Vec<Spec>> where
T: AsRef<str>,

Get a valid, sorted Vec<Spec> given a Vec<String>.

Accounts for the following error conditions:

volta install node 12, where the user intended to install node@12 but used syntax like in nodenv or nvm

invalid version specs

Returns a listed sorted so that if node is included in the list, it is always first.

fn check_args<T>(args: &[T], action: &str) -> Fallible<()> where
T: AsRef<str>,

Check the args for the bad pattern of volta install <tool> <number>.

fn sort_comparator(left: &Spec, right: &Spec) -> Ordering

Compare Specs for sorting when converting from strings

We want to preserve the original order as much as possible, so we treat tools in the same tool category as equal. We still need to pull Node to the front of the list, followed by Npm / Yarn, and then Packages last.

impl Spec

pub fn resolve(self, session: &mut Session) -> Fallible<Resolved>

Resolve a tool spec into a fully realized Tool that can be fetched

pub fn uninstall(self) -> Fallible<()>

Uninstall a tool, removing it from the local inventory

This is implemented on Spec, instead of Resolved, because there is currently no need to resolve the specific version before uninstalling a tool.