Module volta_core::tool::node

Modules

fetch

Provides fetcher for Node distributions

metadata

resolve

Provides resolution of Node requirements into specific versions, using the NodeJS index

Structs

Node

The Tool implementation for fetching and installing Node

NodeVersion

A full Node version including not just the version of Node itself but also the specific version of npm installed globally with that Node installation.

Constants

NODE_DISTRO_ARCH

The architecture component of a Node distro filename

NODE_DISTRO_EXTENSION

The extension for Node distro files

NODE_DISTRO_IDENTIFIER

The file identifier in the Node index files array

NODE_DISTRO_OS

The OS component of a Node distro filename

Functions

load_default_npm_version

Load the local npm version file to determine the default npm version for a given version of Node

resolve