Module volta_core::tool::node::fetch

Provides fetcher for Node distributions

Structs

Manifest 🔒

The portion of npm’s package.json file that we care about

Functions

determine_remote_url 🔒

Determine the remote URL to download from, using the hooks if available

fetch

fetch_remote_distro 🔒

Fetch the distro archive from the internet

load_cached_distro 🔒

Return the archive if it is valid. It may have been corrupted or interrupted in the middle of downloading.

load_default_npm_version

Load the local npm version file to determine the default npm version for a given version of Node

npm_manifest_path 🔒

public_node_server_root 🔒

save_default_npm_version 🔒

Save the default npm version to the filesystem for a given version of Node

unpack_archive 🔒

Unpack the node archive into the image directory so that it is ready for use