Module volta_core::tool::package

Re-exports

pub use install::BinConfig;

pub use install::BinLoader;

pub use install::PackageConfig;

pub use resolve::resolve;

Modules

fetch
Provides fetcher for 3rd-party packages

install
Provides installation steps for 3rd-party packages, fetching their dependencies, writing config files, and creating shims

resolve
Provides resolution of 3rd-party packages into specific versions, using the npm repository

serial

Structs

Package
The Tool implementation for fetching and installing 3rd-party packages

PackageDetails
Details required for fetching a 3rd-party Package

Functions

bin_full_path

binaries_from_package
Reads the contents of a directory and returns a Vec containing the names of all the binaries installed by the input package.

remove_config_and_shim

uninstall
Uninstall the specified package.