Module volta_core::tool::package::manager

Enums

PackageManager
The package manager used to install a given package

Constants

_IMPL_DESERIALIZE_FOR_PackageManager

_IMPL_SERIALIZE_FOR_PackageManager

Functions

get_npm_package_name
Determine the package name for an npm global install

get_single_directory_name
Return the name of the single subdirectory (if any) to the given parent_dir

get_yarn_package_name
Determine the package name for a Yarn global install