Trait volta_core::tool::Tool

pub trait Tool: Display {
    fn fetch(self, session: &mut Session) -> Fallible<()>;

    fn install(self, session: &mut Session) -> Fallible<()>;

    fn pin(self, session: &mut Session) -> Fallible<()>;
}

Trait representing all of the actions that can be taken with a tool

Required methods

fn fetch(self, session: &mut Session) -> Fallible<()>

Fetch a Tool into the local inventory

fn install(self, session: &mut Session) -> Fallible<()>

Install a tool, making it the default so it is available everywhere on the user's machine

fn pin(self, session: &mut Session) -> Fallible<()>

Pin a tool in the local project so that it is usable within the project

Loading content...